// plugins/plugin-chart-leaflet/src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import {
  QueryFormData,
  TimeseriesDataRecord,
  ChartProps
} from '@superset-ui/core';
import { GeoJsonObject, Feature } from 'geojson';

export interface SupersetPluginChartLeafletStylesProps {
  height?: number;
  width?: number;
  headerFontSize?: string;
  boldText?: boolean;
}

export type SupersetPluginChartLeafletQueryFormData = QueryFormData &
  SupersetPluginChartLeafletStylesProps & Record<string, any>;

export type SupersetPluginChartLeafletProps = SupersetPluginChartLeafletStylesProps &
    ChartProps &
  {
    data: TimeseriesDataRecord[];
    payload: any;
    formData: SupersetPluginChartLeafletQueryFormData;
    // add typing here for the props you pass in from transformProps.ts!
  };

export type Range = [number, number];
export type Point = [number, number];
export interface ColorType {
  r: number;
  g: number;
  b: number;
  a: number;
}
export interface ColorSetup {
  dataType: string;
  colorLimits: number[];
  colorMap: Record<string, string>;
  colorScale: any | null;
  colorDataType: string;
  legendColorName: string;
  colorData: any[];
  sortedColorData: any[];
  legendTitleSize?: number | string;
}

export type GeoJSONFeature = Feature | GeoJsonObject | any;
export interface LeafletSetup {
  chartData: Record<string, TimeseriesDataRecord[]>;
  chartParams: Record<string, SupersetPluginChartLeafletQueryFormData>;
  colNames: Record<string, string[]>;
  colTypes: Record<string, string[]>;
  colSetup: Record<string, ColorSetup>;
  chartQueryContext: Record<string, any>;
}

export default ColorSetup;